WebLogin: Clean up no-cookie testing
print_headers was previously returning the result of header_props,
which is defined to return the hash of properties.  This could result
in bizarre page text when the return value was used as the page body,
as it was when redirecting to test cookie handling.

Change print_headers to always return the empty string, and change
all of the error probing done at the start of login to check whether
the return value is defined rather than true, allowing those functions
to return the empty string to send the browser an empty page (generally
with redirect headers).

Change-Id: I5cf99978d6ac620ca5f803c76d7b4a5309bc7838
Reviewed-on: https://gerrit.stanford.edu/1182
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>