WebLogin: Fix remember_login setting for login page
When using the default direction of remember_login (a checkbox
that has to be set to remember the login), we can't just send the
query parameter if present.  We have to send to the template whatever
the current value is, so that the template can make the correct
decision for how to show both cases.

Change-Id: Ie2d916eed3611ed54e04b6036878b4b22c10f13e
Reviewed-on: https://gerrit.stanford.edu/1195
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>