Fix WebAuthOptional with Apache 2.4
WebAuthOptional was previously broken with Apache 2.4 because the
behavior of require valid-user changed to require a username be set.
For Apache 2.4 and later, set the authenticated user to the empty
string when WebAuthOptional is enabled and the user isn't
authenticated, which restores the behavior of Apache 2.2.  Be aware
that this means REMOTE_USER may show up in the environment when
WebAuthOptional is set the user is not authenticated, but the value
will be empty.

Change-Id: Iff9a2818b508c3a7bc5808772ebfa250b5c48072
Reviewed-on: https://gerrit.stanford.edu/1215
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>