Add additional sanity checking of WebKDC reply
We saw the following set of errors in production:

    token type not present in WebKDC response; mod_webkdc on the
      WebKDC may be older than 3.6.1
    Use of uninitialized value in concatenation (.) or string at
      /usr/share/perl5/WebLogin.pm line 730.

which appear to indicate that the WebKDC::make_request_token_request
call returned success but didn't include a response token.  It's not
at all clear how that could happen, and the mod_webkdc log looked
normal.  Add some additional sanity checking so that we don't blindly
try to construct a return page with no meaningful response token in
case this error happens again.

Change-Id: Ie7b10cd2fe8bebd00f34c870d9371fc31efa9fab
Reviewed-on: https://gerrit.stanford.edu/1244
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>