Diagnose a missing Kerberos authenticator earlier in login
If a password authentication is required in order to obtain a Kerberos
authenticator, return that error in preference to a multifactor
required error.  This ensures that the password authentication page
happens first, preserving expected user page flow, and fixes various
errors and loops caused by detecting this problem after the successful
second factor authentication.

Change-Id: I6c0d46407bbe95fdaa97e767423d737645f76618
Reviewed-on: https://gerrit.stanford.edu/1248
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>