Log and ignore mismatched webkdc-proxy tokens
If the user presents a login token for one user and a webkdc-proxy
token for a different user, or, more generally, mismatched
webkdc-proxy tokens, ignore and log the mismatched webkdc-proxy token
rather than rejecting the authentication with a fatal error.  While
this case ideally should not happen, in practice it's not uncommon for
users sharing devices to attempt authentication (due to session factor
requirements or forced login) while still possessing webkdc-proxy
tokens for another user, and rejecting the authentication instead of
replacing the older webkdc-proxy token does nothing to improve the
situation.

Change-Id: Icb2dc257e41679f305c2eaf80f847d6e7abcb68d
Reviewed-on: https://gerrit.stanford.edu/1250
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>