Retry WebLogin post to WebKDC, produce better errors
If the WebLogin post to the WebKDC fails, retry once.  It's common for
the POST to be interrupted by a signal from the FastCGI process
manager trying to shut down the login.fcgi process, in which case
retrying will succeed and allow completion of the request before
shutting down.

Produce more succinct and hopefully still useful error messages when
WebLogin cannot POST to the WebKDC.

Change-Id: Id0173e7b079d6f5169107cc3465b56591373d3ca
Reviewed-on: https://gerrit.stanford.edu/1251
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>