Handle invalid XML from the WebKDC in WebLogin
Work around problems with WebLogin parsing of the XML returned from
the WebKDC when a user attempts an authentication using a non-ASCII
principal name.  This results in invalid XML that XML::Parser cannot
parse.  The proper fix is to catch this on the WebKDC side, but, as an
interim measure, replace non-ASCII characters in the WebKDC reply with
periods so that reply processing can continue.

Improve error reporting of unparsable XML received by the WebLogin
server from the WebKDC.

Change-Id: I42f9d141c56cae58dfa54a9d8a8d2350a3488f9e
Reviewed-on: https://gerrit.stanford.edu/1271
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>