Don't restart the SMS multifactor process on an error
Support remembering that the user has been sent an SMS message already
when redisplaying the multifactor login page after an error.  For this
to work properly, local templates will have to be updated to set the
form parameter multifactor_sentauth if an SMS message has already been
sent.  See the sample multifactor.tmpl file for an example.

Change-Id: I1f023a2c24e22300a9cac680432088a0f118a002
Reviewed-on: https://gerrit.stanford.edu/1288
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>