Use r->request_config for mod_webauth's context
Using the per-request config allows mod_webauth to persist context
between hook functions.  Because the context keeps references to
the directory config and server config, set those as soon in the
request as we can (as soon as we have a per_dir_config).  In the
httpd 2.2 server this is done in access_checker; in the httpd 2.4
server this is moved to post_perdir_config.

Change-Id: I8fbfa7bb54c169dbc83d18ee26fc494a5d0e884e
Reviewed-on: https://gerrit.stanford.edu/1333
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>