Don't remove request notes in mod_webauth
I found gather_tokens returns 302 in check_user_id in the subrequest
if WebAuthExtraRedirect is on, so check_user_id never gets to set the
N_SUBJECT note, but on the second pass through check_user_id (back in
the main request) the N_WEBAUTHR and N_WEBAUTHS notes are missing, so
mod_webauth thinks we need to go back to WebLogin to get them.

This can be fixed by not removing N_WEBAUTHR and N_WEBAUTHS from the
main request's notes.  Then, subrequests should behave in the same way
as the main request.

Change-Id: I2895ffa042291d8fe78ad4b53ddad51ceb2bfc9e
Reviewed-on: https://gerrit.stanford.edu/1337
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>