Add support for Kerberos password change via remctl
The WebAuth Kerberos API now supports Kerberos password change via the
remctl protocol, which is more robust than the kpasswd protocol when
password changes can take some time.  This can be configured via the
new webauth_krb5_change_config function.  The remote remctl server
must provide a command and subcommand that takes a single argument,
the new password, and changes the password for the authenticated
principal that sent the command.

Change-Id: I5747eade47032da1aaa194d8fd2caf6a471cad5a
Reviewed-on: https://gerrit.stanford.edu/1443
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>