Fix handling of non-directive sections in module manuals
Based on the mod_fcgid documentation, use the correct method of
labeling non-directive sections so that they get proper sidebar
links.  Remove the code from the clean-apache-manual script that
was cobbling this together.

Set the module status to External instead of Contributed, and add
a compatibility section to each module documentation page.

Change-Id: I1122b607d06be1f55328e3064820b0b1010c6a1e
Reviewed-on: https://gerrit.stanford.edu/1454
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>