Ensure the keyring can be loaded at module entry points
At each module entry point that might perform actions with the
keyring, ensure that the keyring is loaded and return an appropriate
error immediately if it's not.  Ensure there are sanity checks in
place for all places the keyring might be used.

Return HTTP_INTERNAL_SERVER_ERROR if configuring the WebKDC fails.

Change-Id: I66ca283e0d6aeb04a759519897939ba2625b2465
Reviewed-on: https://gerrit.stanford.edu/1467
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>