Restructure and improve the mod_webauth tests
Move the logout script up a level so that it isn't covered by the
authentication requirement (auth/logout was weird).  Adjust the
test harness so that tests can use an alternative logout path.
Move the tests for cookie path scoping to a separate directory so
that the whole directory can have the same path scope and they can
have their own logout script.  Simplify the structure of those
tests somewhat.

Change-Id: I75b6875886730442d4984bc25f35640ba1d14448
Reviewed-on: https://gerrit.stanford.edu/1473
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>