Improve WebLogin logic for showing password expiration warning
Show the expiring password warning in WebLogin if the browser request
was a POST.  Previously, it was skipped if the user had a REMOTE_USER
preference or if the browser presented a single sign-on cookie.  This
was too conservative, not warning in cases when REMOTE_USER failed,
when the browser presented an expired single sign-on cookie (systems
that are suspended rather than shut down, for example), and when the
user has to do multifactor authentication.  Checking for a POST is a
closer match for when we can force a confirmation screen without too
much user disruption.

Change-Id: Iacfe1b3d5f5962fe303157e6d9f8ce6be47f33eb
Reviewed-on: https://gerrit.stanford.edu/1483
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>