Build correctly when remctl support is disabled
The new remctl-based password change protocol broke the build of
the library when remctl support was not enabled due to an
incorrectly-named stub function.  Fix the function name and
diagnose attempting to configure remctl-based password change
without support for it earlier in the code path.

Thanks to Peter Wienemann for the report.

Change-Id: I120418bb8aaff7c6c406ef6965fcdd5e6ab1900d
Reviewed-on: https://gerrit.stanford.edu/1486
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>