Fix various grammar and wording issues in the protocol spec
Clarify the contents of the token returned to the WAS from the
WebKDC and the reason for having the session key both outside and
inside the encrypted token.  Fix various other grammar and wording
mistakes, including using a more appropriate preposition than "in"
for specifying the key used for an encryption.

Thanks to Adam Lewenberg for the review.

Change-Id: I3cdea8b61c5029684b9cd75140f113a23d3e5806
Reviewed-on: https://gerrit.stanford.edu/1537
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>