Add FAST support, fix WebKdcKerberosFactors
Add a new configuration directive, WebKdcFastArmorCache, for
mod_webkdc.  If set, this specifies the path to a Kerberos ticket
cache that can (and must) be used for FAST (Flexible Authentication
Secure Tunneling) protection of Kerberos password authentications.
The Kerberos KDC must also support FAST in order to safely enable this
option.  Based on a patch by Jakob Uhd Jepsen (One.com A/S).

Fix parsing of the WebKdcKerberosFactors configuration directive.

Change-Id: Ib2849b267e83e07d376eb4ab2dd64db75cb1e7b7
Reviewed-on: https://gerrit.stanford.edu/1540
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>