Translate KRB5_KDC_UNREACH to WA_PEC_USER_REJECTED
When translating Kerberos errors, treat KRB5_KDC_UNREACH (cannot
contact any KDC for realm) as a user rejected error instead of a
Kerberos error.  This avoids returning an internal error from WebLogin
and instead tells the user the username is invalid.  This is not
always correct, since the unreachable KDC could be the local KDC, but
it's better than the previous behavior of throwing internal errors
when users enter email addresses as their username.

Change-Id: Ie9d60370d6dc0e6878f046301e467d1965a7412b
Reviewed-on: https://gerrit.stanford.edu/1544
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>