Bump shared library versioning
The change in the webauth_webkdc_config struct means that some
interfaces are no longer compatible with previous releases.  Be
conservative and bump all of the library versioning.  (Normally
I wouldn't do an ABI bump in a minor release, but in this case
the effect of the changes is very minor, just still an ABI break.)

Change-Id: Ibbb515d7d9393f63751f11d05e5a135c1e3aa5dd
Reviewed-on: https://gerrit.stanford.edu/1579
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>