Add new device_id field to login tokens
This will be used by WebLogin to pass the authenticating device
identifier through to the WebKDC.  Add the field to the login
token encoding and to all of the various tests, and adjust the
test suite construction of login tokens to account for the extra
field.  Pass through a value in multifactor tests in preparation
for further tests that actually use that field.

Change-Id: I00cfbe3ed3ebe57294c8caa4c0e1041301e21577
Reviewed-on: https://gerrit.stanford.edu/1607
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>