Add support for default and device info in userinfo calls
In the JSON implementation of the user information call, add
support for reading default device information and the list of
configured devices.  Add support for passing in the device ID
for validate calls, and pull it out of the login token.  (This
field is ignored with the old XML protocol.)

Change-Id: Iffadd85499024000b459dada843de29a1b9462ed
Reviewed-on: https://gerrit.stanford.edu/1610
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>