Give a validation remctl command its own timeout error
In order to set up out of band methods that will time out if the user
doesn't respond to the out of band prompt (such as a phone call), we
need to have a validation remctl call tell us specifically if it timed
out.  Currently it's just sent as a WK_ERR_UNRECOVERABLE_ERROR, which
makes it impossible to carve out special logic for a timeout.  This will
now be sent as its own error code.

Currently this is only true for validate and not for the userinfo
command.  If we later need to do special things for userinfo timeouts as
well, we'll build on this work.